+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ## [Unreleased]
 
+ - Added the ability to delete your groups.
  - Cleaned up some UI issues on the profile page.
  - We made the mute button a bit more noticeable.
  - Moved the group creation button, uncluttering your home screen!